# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: meow~traffic API
  description: Backend API for the meow~traffic network traffic analysis platform.
  version: 0.0.1
  contact:
    name: meow~traffic
servers:
  - url: http://127.0.0.1:17891
    description: Local backend
security:
  - BearerAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
  parameters:
...
    StreamId:
      name: streamId
      in: query
      required: true
      schema:
        type: integer
        format: int64
      description: Stream ID
...
  schemas:
...
    StreamChunk:
      type: object
      properties:
        packet_id:
          type: integer
          format: int64
        direction:
          type: string
        body:
          type: string
    ReassembledStream:
      type: object
      properties:
        stream_id:
          type: integer
          format: int64
        protocol:
          type: string
        from:
          type: string
        to:
          type: string
        chunks:
          type: array
          items:
            $ref: '#/components/schemas/StreamChunk'
        request:
          type: string
        response:
          type: string
        load_meta:
          type: object
          properties:
            source:
              type: string
            loading:
              type: boolean
            cache_hit:
              type: boolean
            index_hit:
              type: boolean
            file_fallback:
              type: boolean
            tshark_ms:
              type: integer
              format: int64
            override_count:
              type: integer
...
    ToolRuntimeSnapshot:
      type: object
      properties:
        config:
          type: object
        tshark:
          type: object
          properties:
            path:
              type: string
            version:
              type: string
            available:
              type: boolean
            custom_path:
              type: boolean
        ffmpeg:
          type: object
          properties:
            path:
              type: string
            version:
              type: string
            available:
              type: boolean
    TLSConfig:
      type: object
      properties:
        enabled:
          type: boolean
        cert_file:
          type: string
        key_file:
          type: string
        min_version:
          type: string
    MCPConfig:
      type: object
      properties:
        enabled:
          type: boolean
        transport:
          type: string
        endpoint:
          type: string
    MCPStatus:
      type: object
      properties:
        enabled:
          type: boolean
        transport:
          type: string
        endpoint:
          type: string
        auth_required:
          type: boolean
...
paths:
...
  /api/streams/http:
    get:
      tags: [Streams]
      summary: Get HTTP stream
      description: Returns an HTTP stream reassembled into request/response pairs.
      operationId: getHTTPStream
      parameters:
        - $ref: '#/components/parameters/StreamId'
      responses:
        '200':
          description: HTTP stream data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReassembledStream'
...
  /api/tools/runtime-config:
    get:
      tags: [Tools]
      summary: Get runtime config
      description: Returns the full tool runtime configuration snapshot.
      operationId: getToolRuntimeConfig
      parameters:
        - name: probe
          in: query
          schema:
            type: string
            default: full
          description: Probe mode (full, skip, etc.)
      responses:
        '200':
          description: Runtime configuration snapshot
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ToolRuntimeSnapshot'
    post:
      tags: [Tools]
      summary: Update runtime config
      description: Updates the tool runtime configuration (paths, options).
      operationId: setToolRuntimeConfig
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                tshark_path:
                  type: string
                python_path:
                  type: string
                ffmpeg_path:
                  type: string
      responses:
        '200':
          description: Updated runtime configuration
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ToolRuntimeSnapshot'
...
  /api/tls:
    get:
      tags: [TLS]
      summary: Get TLS config
      description: Returns the current TLS configuration.
      operationId: getTLSConfig
      responses:
        '200':
          description: TLS configuration
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TLSConfig'
    post:
      tags: [TLS]
      summary: Update TLS config
      description: Updates the TLS configuration (cert, key, min version).
      operationId: setTLSConfig
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TLSConfig'
      responses:
        '200':
          description: TLS config updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: updated
  /api/mcp/config:
    get:
      tags: [MCP]
      summary: Get MCP config
      description: Returns the current MCP server configuration and status.
      operationId: getMCPConfig
      responses:
        '200':
          description: MCP status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MCPStatus'
    post:
      tags: [MCP]
      summary: Update MCP config
      description: Updates the MCP server configuration.
      operationId: setMCPConfig
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MCPConfig'
      responses:
        '200':
          description: Updated MCP status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MCPStatus'
...